requirements.txt
langchain-ibm
langchain-community
pdfplumber
python-dotenv
streamlit
langgraph